requirements.txt
requests==2.32.3
python-dotenv==1.1.0
python-telegram-bot==22.0